server/bedrock/chat/repository.py
"""Repository for chat messages in DynamoDB.

This module provides a data access layer for storing and retrieving chat messages
from a DynamoDB table. It follows the existing service patterns for error handling
and logging to maintain consistency with the main application.

STUDENT IMPLEMENTATION REQUIRED:
- save_message(): Implement DynamoDB message persistence
- get_messages(): Implement DynamoDB message retrieval

Both methods include detailed TODO comments with implementation guidance.
"""

import logging
import time
from typing import Dict, List, Optional, Any

import boto3
from boto3.dynamodb.conditions import Key
from botocore.exceptions import ClientError

from .exceptions import (
    ChatServiceUnavailableError,
    ChatSessionNotFoundError,
    ChatError,
)

logger = logging.getLogger(__name__)


class MessageRepository:
    """Repository for chat messages in DynamoDB.

    This class provides methods for storing and retrieving chat messages from a DynamoDB table.
    The table schema is as follows:
    - Partition Key: session_id (String) - Unique identifier for each conversation session
    - Sort Key: message_id (String) - Unique identifier for each message within a session
    """

    def __init__(self, table_name: str, dynamodb_client=None):
        """Initialize the repository with the DynamoDB table name.

        Args:
            table_name: The name of the DynamoDB table to use
            dynamodb_client: Optional DynamoDB client for testing
        """
        self.table_name = table_name
        self.dynamodb = dynamodb_client or boto3.resource("dynamodb")
        self.table = self.dynamodb.Table(table_name)

    def save_message(
        self, session_id: str, message_id: str, message_type: str, message: str,
        intent: Optional[str] = None, sources: Optional[List[Dict[str, Any]]] = None
    ) -> Dict[str, Any]:
        """Save a message to the DynamoDB table.

        Args:
            session_id: The session ID
            message_id: The message ID
            message_type: The message type (HUMAN or BOT)
            message: The message content
            intent: The message intent (CHAT or QUERY)
            sources: List of source documents used for RAG responses

        Returns:
            The saved message item
            
        Raises:
            ChatServiceUnavailableError: If the DynamoDB service is unavailable
            ChatError: For other DynamoDB errors
        """
        # TODO: Implement message saving to DynamoDB
        #
        # STUDENT IMPLEMENTATION REQUIRED:
        # This method needs to save a chat message to the DynamoDB table.
        # Follow the existing service patterns for error handling and logging.
        #
        # Implementation steps:
        # 1. Create the item dictionary with all required fields:
        #    - session_id, message_id, message_type, message (required)
        #    - timestamp: use int(time.time())
        #    - intent, sources: include only if provided (not None)
        #
        # 2. Use self.table.put_item(Item=item) to save to DynamoDB
        #
        # 3. Handle ClientError exceptions following this pattern:
        #    - ServiceUnavailable, InternalServerError, ThrottlingException 
        #      -> raise ChatServiceUnavailableError("DynamoDB")
        #    - ProvisionedThroughputExceededException 
        #      -> raise ChatServiceUnavailableError("DynamoDB throughput exceeded")
        #    - Other ClientError -> raise ChatError(f"DynamoDB error: {str(e)}", 500)
        #
        # 4. Add logging: logger.info(f"Saved message {message_id} for session {session_id}")
        #
        # 5. Return the complete item dictionary
        #
        # Hint: Look at the existing service's DynamoDB patterns for reference
        
        raise NotImplementedError("Students need to implement DynamoDB message saving")

    def get_messages(self, session_id: str) -> List[Dict[str, Any]]:
        """Get all messages for a session.

        Args:
            session_id: The session ID

        Returns:
            A list of message items
            
        Raises:
            ChatServiceUnavailableError: If the DynamoDB service is unavailable
            ChatError: For other DynamoDB errors
        """
        # TODO: Implement message retrieval from DynamoDB
        #
        # STUDENT IMPLEMENTATION REQUIRED:
        # This method needs to retrieve all messages for a session from DynamoDB.
        # Follow the existing service patterns for error handling and logging.
        #
        # Implementation steps:
        # 1. Use self.table.query() with these parameters:
        #    - KeyConditionExpression=Key("session_id").eq(session_id)
        #    - ScanIndexForward=True  # For chronological order (oldest first)
        #
        # 2. Extract items from response: items = response.get("Items", [])
        #
        # 3. Handle ClientError exceptions (same pattern as save_message):
        #    - ServiceUnavailable, InternalServerError, ThrottlingException 
        #      -> raise ChatServiceUnavailableError("DynamoDB")
        #    - ProvisionedThroughputExceededException 
        #      -> raise ChatServiceUnavailableError("DynamoDB throughput exceeded")
        #    - Other ClientError -> raise ChatError(f"DynamoDB error: {str(e)}", 500)
        #
        # 4. Add logging: logger.info(f"Retrieved {len(items)} messages for session {session_id}")
        #
        # 5. Return the list of message items
        #
        # Hint: Import Key from boto3.dynamodb.conditions at the top of the file
        
        raise NotImplementedError("Students need to implement DynamoDB message retrieval")

