server/bedrock/chat/requirements.txt
# Core dependencies
boto3>=1.34.0
langchain>=0.1.0
langchain-aws>=0.1.0
langchain-core>=0.1.0
pydantic>=2.0.0
ulid-py>=1.1.0

# AWS Lambda Powertools (for metrics and logging)
aws-lambda-powertools>=2.0.0

# Optional: For local development
python-dotenv>=1.0.0
moto>=5.0.0  # For mocking AWS services in tests
